requirements.txt
Appium-Python-Client==0.36
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
nose==1.3.7
pluggy==0.6.0
py==1.5.3
pytest==3.6.3
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-ordering==0.5
requests==2.19.0
selenium==3.141.0
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1

